requirements.txt
streamlit 
pandas 
phidata
groq
googlesearch-python
pycountry
yfinance
dotenv
